Refactor user-related files to a new folder
